Extrage logica de import factură într-un mixin reutilizabil (port din 18.0)
Portare a refactorului din 18.0 (dhongu/deltatech#2624): extrage matching-ul
de produse, actualizarea prețului furnizor, crearea/actualizarea liniilor
de comandă, validarea recepției și crearea facturii furnizor din
purchase.ubl.import.wizard într-un model abstract nou,
purchase.invoice.import.mixin. Wizard-ul UBL îl folosește prin _inherit
și își păstrează neschimbate părțile specifice XML.

Portarea a păstrat adaptările specifice 19.0 deja prezente în acest branch
(product_uom_id în loc de product_uom pe liniile de comandă, move_ids în
loc de move_ids_without_package eliminat, _set_quantity_done/picked în loc
de qty_done) — nu e o copie mecanică a fișierului din 18.0, ci același
refactor aplicat pe codul specific 19.0. Toate cele 10 teste existente
trec neschimbate.

Permite altor wizard-uri de import (facturi PDF Marso/Delta/Sigemo/Procar,
portate separat) să reutilizeze aceeași logică de procesare fără
duplicare.
