+++ deltatech_purchase_ubl/__init__.py
@@ -2,5 +2,5 @@
 # Dorin Hongu <dhongu(@)gmail(.)com>
 # See README.rst file on addons root folder for license details
 
-from . import wizard
 from . import models
+from . import wizard
