+++ deltatech_purchase_ubl/models/__init__.py
@@ -2,4 +2,5 @@
 # Dorin Hongu <dhongu(@)gmail(.)com>
 # See README.rst file on addons root folder for license details
 
+from . import purchase_invoice_import_mixin
 from . import purchase_order
